Make gallery index more generic
Changes to make the index page for a gallery can
be made as specific as possible.

Includes addition of a plugin to let you include
page variables in frontmatter. Approach taken
from:

http://acegik.net/blog/ruby/jekyll/plugins/howto-nest-liquid-template-variables-inside-yaml-front-matter-block.html

Also, we're moving some data up the inheritance
chain into layouts, we need to refer to it in the
liquid with a `layout` prefix, so `layout.title`
rather than `page.title`. This means I've ended up
making new versions of the base and gallery
layouts for only the covid-year gallery to use, so
these changes don't break any existing pages.